resolved error

+++ app/src/main/java/com/example/xpensate/API/BudgetBuilder/BudgetExpensesResponse.kt
@@ -1,10 +1,10 @@
 package com.example.xpensate.API.BudgetBuilder
 
 data class BudgetExpensesResponse(
-    val luxury_credit_total: Int,
-    val luxury_debit_total: Int,
-    val needs_credit_total: Int,
-    val needs_debit_total: Int,
+    val luxury_credit_total: Double,
+    val luxury_debit_total: Double,
+    val needs_credit_total: Double,
+    val needs_debit_total: Double,
     val success: Boolean,
     val monthly: Int
 )
