+++ app/src/main/java/com/example/xpensate/API/home/IncomevsExpense.kt
@@ -2,6 +2,6 @@ package com.example.xpensate.API.home
 
 data class IncomevsExpense(
     val success: Boolean,
-    val total_expense: Int,
-    val total_income: Int
+    val total_expense: Double,
+    val total_income: Double
 )
